Implement webroot solver in addition to standalone
Since cPanel DNSONLY 86, it seems that, at least some of the time, that
cpsrvd is bound to port 80 and 443. Luckily, the HTTP handler for port
80 appears to check for static content in /usr/local/apache/htdocs/, so
we can drop our challenge response files in there.

With this commit, fleetssl-dnsonly attempts to both the standalone and
webroot challenges - one of them should succeed.

Fixes #15