Add workaround for bugged /usr/local/apache perms
According to reports in #16, sometimes some of directories leading up to
the Apache webroot are not world readable, thus preventing the challenge
file from being accessible. I can only guess that this is some kind of a
bug in the DNSONLY installer/upgrader.

This commit adds a workaround by upgrading go+rx permissions on every
directory in the tree leading to the webroot, if those permission bits
are not set.

Fixes #16.